make refreshIdToken reusable (hook)
